Preparing to publish

+++ src/bindgen/constants/IPV4_HDR.rs
@@ -2,11 +2,11 @@
 // Copyright © 2016 The developers of dpdk-sys. See the COPYRIGHT file in the top-level directory of this distribution and at https://raw.githubusercontent.com/lemonrock/dpdk-sys/master/COPYRIGHT.
 
 
-pub const IPV4_HDR_DF_FLAG: c_int = 16384;
-pub const IPV4_HDR_DF_SHIFT: c_int = 14;
-pub const IPV4_HDR_FO_SHIFT: c_int = 3;
-pub const IPV4_HDR_IHL_MASK: c_int = 15;
-pub const IPV4_HDR_MF_FLAG: c_int = 8192;
-pub const IPV4_HDR_MF_SHIFT: c_int = 13;
-pub const IPV4_HDR_OFFSET_MASK: c_int = 8191;
-pub const IPV4_HDR_OFFSET_UNITS: c_int = 8;
+pub const IPV4_HDR_DF_FLAG: uint16_t = 16384;
+pub const IPV4_HDR_DF_SHIFT: usize = 14;
+pub const IPV4_HDR_FO_SHIFT: usize = 3;
+pub const IPV4_HDR_IHL_MASK: uint8_t = 15;
+pub const IPV4_HDR_MF_FLAG: uint16_t = 8192;
+pub const IPV4_HDR_MF_SHIFT: usize = 13;
+pub const IPV4_HDR_OFFSET_MASK: uint16_t = 8191;
+pub const IPV4_HDR_OFFSET_UNITS: usize = 8;
